requirements.txt
aenum==3.0.0
aiofiles==0.6.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
blessed==1.18.0
boto3==1.17.24
botocore==1.20.24
bullet==2.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorclass==2.2.0
dacite==1.6.0
dataclass-csv==1.2.0
dnspython==2.1.0
docopt==0.6.2
ecdsa==0.14.1
email-validator==1.1.2
enlighten==1.8.0
Events==0.4
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
halo==0.0.31
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
log-symbols==0.0.14
lxml==4.6.2
MarkupSafe==1.1.1
Naked==0.1.31
orjson==3.5.1
packaging==20.9
pip-upgrader==1.4.15
promise==2.3
py-getch==1.0.1
pydantic==1.8.1
pyfiglet==0.8.post1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
PyYAML==5.4.1
rapidfuzz==1.2.1
requests==2.25.1
Rx==1.6.1
s3transfer==0.3.4
six==1.15.0
spinners==0.0.24
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
tabulate==0.8.9
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.59.0
ujson==4.0.2
urllib3==1.26.3
uvicorn==0.13.4
uvloop==0.15.2
vigorish==0.4.6
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1